disable a test in multi function in join
With postgres >= 13, this query is router executable:

SELECT * FROM table1 JOIN add(3,5) sum ON (id = sum) ORDER BY id ASC;

But the function add(int,int) doesn't exist in workers so it errors out.
Disabling it is probably not a good idea and we should understand why it
is router executable when it is not with pg < 13.