Fix running VACUUM twice if the local node has metadata
If VACUUM is run on a node that has metadata, then VACUUM would be run
twice because we first call it in standardUtility and then we call it in
PostprocessVacuumStmt. Now while creating vacuum tasks we only consider
remote nodes so that we don't run it twice. This would get it stuck in
pg13 because with pg13 VACUUM is restricted as one backend per database.