VACUUM local shard placements with shell table
When a utility command is given, we would use the standard_utility to
process the shell table and then we might process the shards as well.
For VACUUM we would do it like that but postgres 13 started to take
global locks on VACUUM so that there is only one VACUUM per database.

The problem with this is the following:
- We execute the VACUUM for the shell table through the standard process
utility. In this step the global lock is taken for the current database.
- If the current node has shard placements then it tries to execute
VACUUM over a connection to localhost with ExecuteUtilityTaskList.
- the VACUUM on shard placements cannot proceed because it is waiting
for the global lock for the current database to be released.
- The acquired lock from the VACUUM for shell table will not be released
until the transaction is committed.
- So there is a deadlock.

In order to solve this there are different solutions:

1-> We could skip executing the VACUUM on the shell table as the shell
table only has the data that needs to be truncated. However we cannot
just skip the standard process utility for the shell table, because we
need to take the necessary locks for the shell table so that conflicting
commands cannot proceed in parallel.

We haven't chosen this solution because this requires copy-pasting many
code from the postgres side. In order to take the locks, there are many
iterative steps and copy-pasting all of those is not that practical as
it might get out-dated really easily.

2-> We could commit the transaction after the shell table is done so
that the lock could be released but this pushes the problem to shard
placements. Basically shard placements would wait for the lock. Though
we cannot execute the VACUUM commands in parallel because of the global
lock described above.

3-> We could append the local shard placements to the shell table so
that postgres handles everything for us, this would mean we lose the
parallelism but we already cannot have the parallelism in any case
because of the global lock per database. So basically we do all the
VACUUM for the current node in one go, then we continue propogating the
VACUUM to the workers.