fix stackoverflow for the colocated placement size query
We used to query the colocated placement size by constructing a concatination of + size(relation).
This would cause a stackoverflow when there was a big set of colocated placements.

This patch rewrites the query in a way where it should iterate through a list of relation names and
calculate the sum of the sizes via an aggregate.