Dissuade planner from choosing index-only scan for columnar tables
While we are at it, let's prevent index-only scan for columnar tables
by using `ColumnarGetRelationInfoHook`.

Previously, even when `EXPLAIN` output tells that we will do
index-only scan, it was never the case since columnar tables don't
have the visibility info that postgres is looking for.

For this reason, visibility check done in `IndexOnlyNext->VM_ALL_VISIBLE`
code-path was always returning false and postgres was reading the
tuple from the columnar relation itself.