fixup! Skip aborted writes when reading columnar tables
For sequential / columnar scan, consider aborted writes as unflushed
too. This fixes the bug around REINDEX statements when there are
aborted writes. Interestingly, this only happens if we disabled parallel
query for the table.

For index scan, distiguish aborted and in-progress writes.
Make sure to return false for aborted ones, but for in-progress ones,
return true without filling tupleslot as before.